Fix HMI command parsing for menu item changes and pedalboard save
The Protocol parser required commands to be registered via
register_cmd_callback, but HMI commands were never registered.
Replace with direct string parsing for incoming HMI commands.
Also fix 'c' command arg type from Int to Float for BPM values.

+++ src/mod_protocol.rs
@@ -65,7 +65,7 @@ pub fn cmd_args() -> HashMap<&'static str, HashMap<&'static str, Vec<ArgType>>> 
     all.insert("restore", vec![]);
     all.insert("screenshot", vec![Int, Str]);
     all.insert("r", vec![Int]);
-    all.insert("c", vec![Int, Int]);
+    all.insert("c", vec![Int, Float]);
     all.insert("upr", vec![Int]);
     all.insert("ups", vec![Int]);
     all.insert("lp", vec![Int]);
